requirements.txt
web3==6.0.0
python-dotenv==0.20.0
Flask==2.2.0